+++ extend/waibuMpa/extend/waibuAdmin/route/team-setting/@action.js
@@ -1,6 +1,6 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'manageTeamSetting',
+  title: 'teamSetting',
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')