+++ extend/waibuMpa/extend/waibuAdmin/route/user-setting/@action.js
@@ -1,10 +1,10 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'teamUser',
+  title: 'userSetting',
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'SumbaTeamUser', req, reply)
+    return await crudSkel.call(this, 'SumbaUserSetting', req, reply)
   }
 }