+++ extend/waibuMpa/extend/waibuAdmin/route/user/@action.js
@@ -1,10 +1,11 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'manageUser',
+  title: 'userProfile',
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'SumbaUser', req, reply)
+    const options = { modelOpts: { forceNoHidden: ['token'] } }
+    return await crudSkel.call(this, 'SumbaUser', req, reply, { options })
   }
 }